@@ -17,10 +17,11 @@ find = {}
 test_driver = ["py.typed"]
 
 [tool.ruff]
+target-version = "py312"
 line-length = 88
 
-select = ["E", "F", "I", "U", "N"]
-ignore = ["E501"]
+lint.select = ["E", "F", "I", "U", "N"]
+lint.ignore = ["E501"]
 
 # xxx: we can import https://pypi.org/project/types-colorama/ here
 [[tool.mypy.overrides]]
@@ -31,13 +32,11 @@ ignore_missing_imports = true
 module = "ptpython.*"
 ignore_missing_imports = true
 
-[tool.black]
-line-length = 88
-target-version = ['py39']
-include = '\.pyi?$'
+[[tool.mypy.overrides]]
+module = "junit_xml.*"
+ignore_missing_imports = true
 
 [tool.mypy]
-python_version = "3.10"
 warn_redundant_casts = true
 disallow_untyped_calls = true
 disallow_untyped_defs = true
