@@ -94,5 +94,5 @@ def update_grammars(nvim_treesitter_dir: str):
 
 
 if __name__ == "__main__":
-    generated = update_grammars(sys.args[1])
+    generated = update_grammars(sys.argv[1])
     open(join(os.path.dirname(__file__), "generated.nix"), "w").write(generated)
