@@ -3,7 +3,7 @@
 # run with:
 # $ nix run .\#vimPluginsUpdater
 # format:
-# $ nix run nixpkgs#python3Packages.black -- update.py
+# $ nix run nixpkgs#python3Packages.ruff -- update.py
 # type-check:
 # $ nix run nixpkgs#python3Packages.mypy -- update.py
 # linted:
@@ -19,28 +19,24 @@
 #
 
 import inspect
-import os
-import logging
-import textwrap
 import json
+import logging
+import os
 import subprocess
-from typing import List, Tuple
+import textwrap
 from pathlib import Path
-
+from typing import List, Tuple
 
 log = logging.getLogger("vim-updater")
 
-sh = logging.StreamHandler()
-formatter = logging.Formatter("%(name)s:%(levelname)s: %(message)s")
-sh.setFormatter(formatter)
-log.addHandler(sh)
-
 # Import plugin update library from maintainers/scripts/pluginupdate.py
 ROOT = Path(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))))
-import pluginupdate
 import importlib
-from pluginupdate import run_nix_expr, PluginDesc
-import treesitter
+
+import pluginupdate
+from pluginupdate import PluginDesc, run_nix_expr
+
+treesitter = importlib.import_module("nvim-treesitter.update")
 
 
 HEADER = (
@@ -56,12 +52,32 @@ class VimEditor(pluginupdate.Editor):
     def generate_nix(
         self, plugins: List[Tuple[PluginDesc, pluginupdate.Plugin]], outfile: str
     ):
-        sorted_plugins = sorted(plugins, key=lambda v: v[0].name.lower())
+        log.info("Generating nix code")
+        log.debug("Loading nvim-treesitter revision from nix...")
         nvim_treesitter_rev = pluginupdate.run_nix_expr(
-            "(import <localpkgs> { }).vimPlugins.nvim-treesitter.src.rev", self.nixpkgs
+            "(import <localpkgs> { }).vimPlugins.nvim-treesitter.src.rev",
+            self.nixpkgs,
+            timeout=10,
         )
 
+        GET_PLUGINS_LUA = """
+        with import <localpkgs> {};
+        lib.attrNames lua51Packages"""
+        log.debug("Loading list of lua plugins...")
+        luaPlugins = run_nix_expr(GET_PLUGINS_LUA, self.nixpkgs, timeout=30)
+
+        def _isNeovimPlugin(plug: pluginupdate.Plugin) -> bool:
+            """
+            Whether it's a neovim-only plugin
+            We can check if it's available in lua packages
+            """
+            if plug.normalized_name in luaPlugins:
+                log.debug("%s is a neovim plugin", plug)
+                return True
+            return False
+
         with open(outfile, "w+") as f:
+            log.debug("Writing to %s", outfile)
             f.write(HEADER)
             f.write(
                 textwrap.dedent(
@@ -73,8 +89,8 @@ class VimEditor(pluginupdate.Editor):
                 """
                 )
             )
-            for pdesc, plugin in sorted_plugins:
-                content = self.plugin2nix(pdesc, plugin)
+            for pdesc, plugin in plugins:
+                content = self.plugin2nix(pdesc, plugin, _isNeovimPlugin(plugin))
                 f.write(content)
                 if (
                     plugin.name == "nvim-treesitter"
@@ -84,26 +100,10 @@ class VimEditor(pluginupdate.Editor):
             f.write("\n}\n")
         print(f"updated {outfile}")
 
-    def plugin2nix(self, pdesc: PluginDesc, plugin: pluginupdate.Plugin) -> str:
-        GET_PLUGINS_LUA = """
-        with import <localpkgs> {};
-        lib.attrNames lua51Packages"""
-        luaPlugins = run_nix_expr(GET_PLUGINS_LUA, self.nixpkgs)
-
+    def plugin2nix(
+        self, pdesc: PluginDesc, plugin: pluginupdate.Plugin, isNeovim: bool
+    ) -> str:
         repo = pdesc.repo
-
-        def _isNeovimPlugin(plug: pluginupdate.Plugin) -> bool:
-            """
-            Whether it's a neovim-only plugin
-            We can check if it's available in lua packages
-            """
-            # global luaPlugins
-            if plug.normalized_name in luaPlugins:
-                log.debug("%s is a neovim plugin", plug)
-                return True
-            return False
-
-        isNeovim = _isNeovimPlugin(plugin)
 
         content = f"  {plugin.normalized_name} = "
         src_nix = repo.as_nix(plugin)
@@ -130,19 +130,25 @@ class VimEditor(pluginupdate.Editor):
         if self.nvim_treesitter_updated:
             print("updating nvim-treesitter grammars")
             cmd = [
-                "nix", "build",
-                "vimPlugins.nvim-treesitter.src", "-f", self.nixpkgs
-                , "--print-out-paths"
+                "nix",
+                "build",
+                "vimPlugins.nvim-treesitter.src",
+                "-f",
+                self.nixpkgs,
+                "--print-out-paths",
             ]
             log.debug("Running command: %s", " ".join(cmd))
-            nvim_treesitter_dir = subprocess.check_output(cmd, text=True, timeout=90).strip()
+            nvim_treesitter_dir = subprocess.check_output(
+                cmd, text=True, timeout=90
+            ).strip()
 
             generated = treesitter.update_grammars(nvim_treesitter_dir)
             treesitter_generated_nix_path = os.path.join(
-                NIXPKGS_NVIMTREESITTER_FOLDER,
-                "generated.nix"
+                NIXPKGS_NVIMTREESITTER_FOLDER, "generated.nix"
             )
-            open(os.path.join(args.nixpkgs, treesitter_generated_nix_path), "w").write(generated)
+            open(os.path.join(args.nixpkgs, treesitter_generated_nix_path), "w").write(
+                generated
+            )
 
             if self.nixpkgs_repo:
                 index = self.nixpkgs_repo.index
@@ -159,8 +165,8 @@ class VimEditor(pluginupdate.Editor):
 def main():
     global luaPlugins
 
-    log.debug(f"Loading from {ROOT}/../get-plugins.nix")
-    with open(f"{ROOT}/../get-plugins.nix") as f:
+    log.debug(f"Loading from {ROOT}/get-plugins.nix")
+    with open(f"{ROOT}/get-plugins.nix") as f:
         GET_PLUGINS = f.read()
     editor = VimEditor(
         "vim", Path("pkgs/applications/editors/vim/plugins"), GET_PLUGINS
