@@ -12,9 +12,18 @@ const fixupYarnLock = async (lockContents, verbose) => {
 	const fixedData = Object.fromEntries(
 		Object.entries(lockData.object)
 		.map(([dep, pkg]) => {
+			if (pkg.resolved === undefined) {
+				console.warn(`no resolved URL for package ${dep}`)
+				var maybeFile = dep.split("@", 2)[1]
+				if (maybeFile.startsWith("file:")) {
+					console.log(`Rewriting URL for local file dependency ${dep}`)
+					pkg.resolved = maybeFile
+				}
+				return [dep, pkg]
+			}
 			const [ url, hash ] = pkg.resolved.split("#", 2)
 
-			if (hash || url.startsWith("https://codeload.github.com")) {
+			if (hash || url.startsWith("https://codeload.github.com/")) {
 				if (verbose) console.log(`Removing integrity for git dependency ${dep}`)
 				delete pkg.integrity
 			}
