@@ -7,9 +7,10 @@ import logging
 import subprocess
 import urllib.error
 import urllib.request
+from contextlib import contextmanager
 from operator import itemgetter
 from pathlib import Path
-from typing import List, Dict, Optional, Any, Tuple
+from typing import List, Dict, Optional, Any, Tuple, Set
 
 # We don't want all those deprecated legacy extensions
 # Group extensions by GNOME "major" version for compatibility reasons
@@ -21,6 +22,8 @@ supported_versions = {
     "43": "43",
     "44": "44",
     "45": "45",
+    "46": "46",
+    "47": "47",
 }
 
 # Some type alias to increase readability of complex compound types
@@ -51,9 +54,19 @@ def fetch_extension_data(uuid: str, version: str) -> Tuple[str, str]:
     url: str = f"https://extensions.gnome.org/extension-data/{uuid}.v{version}.shell-extension.zip"
 
     # Download extension and add the zip content to nix-store
-    process = subprocess.run(
-        ["nix-prefetch-url", "--unpack", "--print-path", url], capture_output=True, text=True
-    )
+    for _ in range(0, 10):
+        process = subprocess.run(
+            ["nix-prefetch-url", "--unpack", "--print-path", url], capture_output=True, text=True
+        )
+        if process.returncode == 0:
+            break
+        else:
+            logging.warning(f"Nix-prefetch-url failed for {url}:")
+            logging.warning(f"Stderr: {process.stderr}")
+            logging.warning(f"Retrying")
+
+    if process.returncode != 0:
+        raise Exception("Retried 10 times, but still failed to download the extension. Exiting.")
 
     lines = process.stdout.splitlines()
 
@@ -213,6 +226,20 @@ def process_extension(extension: Dict[str, Any]) -> Optional[Dict[str, Any]]:
     }
 
 
+@contextmanager
+def request(url: str, retries: int = 5, retry_codes: List[int] = [ 500, 502, 503, 504 ]):
+    for attempt in range(retries + 1):
+        try:
+            with urllib.request.urlopen(url) as response:
+                yield response
+                break
+        except urllib.error.HTTPError as e:
+            if e.code in retry_codes and attempt < retries:
+                logging.warning(f"Error while fetching {url}. Retrying: {e}")
+            else:
+                raise e
+
+
 def scrape_extensions_index() -> List[Dict[str, Any]]:
     """
     Scrape the list of extensions by sending search queries to the API. We simply go over it
@@ -226,7 +253,8 @@ def scrape_extensions_index() -> List[Dict[str, Any]]:
         page += 1
         logging.info("Scraping page " + str(page))
         try:
-            with urllib.request.urlopen(
+
+            with request(
                     f"https://extensions.gnome.org/extension-query/?n_per_page=25&page={page}"
             ) as response:
                 data = json.loads(response.read().decode())["extensions"]
@@ -301,15 +329,21 @@ if __name__ == "__main__":
         json.load(out)
 
     with open(updater_dir_path / "collisions.json", "w") as out:
+        # Find the name collisions only for the last 3 shell versions
+        last_3_versions = sorted(supported_versions.keys(), key=lambda v: float(v), reverse=True)[:3]
+        package_name_registry_for_versions = [v for k, v in package_name_registry.items() if k in last_3_versions]
+        # Merge all package names into a single dictionary
+        package_name_registry_filtered: Dict[PackageName, Set[Uuid]] = {}
+        for pkgs in package_name_registry_for_versions:
+            for pname, uuids in pkgs.items():
+                if pname not in package_name_registry_filtered:
+                    package_name_registry_filtered[pname] = set()
+                package_name_registry_filtered[pname].update(uuids)
         # Filter out those that are not duplicates
-        package_name_registry_filtered: Dict[ShellVersion, Dict[PackageName, List[Uuid]]] = {
-            # The outer level keys are shell versions
-            shell_version: {
-                # The inner keys are extension names, with a list of all extensions with that name as value.
-                pname: extensions for pname, extensions in collisions.items() if len(extensions) > 1
-            } for shell_version, collisions in package_name_registry.items()
-        }
-        json.dump(package_name_registry_filtered, out, indent=2, ensure_ascii=False)
+        package_name_registry_filtered = {k: v for k, v in package_name_registry_filtered.items() if len(v) > 1}
+        # Convert set to list
+        collisions: Dict[PackageName, List[Uuid]] = {k: list(v) for k, v in package_name_registry_filtered.items()}
+        json.dump(collisions, out, indent=2, ensure_ascii=False)
         out.write("\n")
 
     logging.info(
