@@ -1,5 +1,7 @@
 #!/usr/bin/env nix-shell
-#!nix-shell -i python3 -p "python3.withPackages (ps: [ ps.beautifulsoup4 ps.lxml ])"
+#!nix-shell -i python3 -p "python3.withPackages (ps: [ ps.beautifulsoup4 ps.lxml ps.packaging ])"
+from functools import cached_property
+from itertools import groupby
 import json
 import os
 import pathlib
@@ -10,6 +12,8 @@ from dataclasses import dataclass
 from enum import Enum
 
 from bs4 import BeautifulSoup, NavigableString, Tag
+from packaging.version import parse as parse_version, Version
+
 
 HERE = pathlib.Path(__file__).parent
 ROOT = HERE.parent.parent.parent.parent
@@ -26,10 +30,22 @@ class KernelNature(Enum):
 class KernelRelease:
     nature: KernelNature
     version: str
-    branch: str
     date: str
     link: str
     eol: bool = False
+
+    @cached_property
+    def parsed_version(self) -> Version:
+        return parse_version(self.version)
+
+    @cached_property
+    def branch(self) -> str:
+        version = self.parsed_version
+        # This is a testing kernel.
+        if version.is_prerelease:
+            return "testing"
+        else:
+            return f"{version.major}.{version.minor}"
 
 
 def parse_release(release: Tag) -> KernelRelease | None:
@@ -37,6 +53,7 @@ def parse_release(release: Tag) -> KernelRelease | None:
     try:
         nature = KernelNature[columns[0].get_text().rstrip(":").upper()]
     except KeyError:
+        # skip linux-next
         return None
 
     version = columns[1].get_text().rstrip(" [EOL]")
@@ -49,21 +66,11 @@ def parse_release(release: Tag) -> KernelRelease | None:
 
     return KernelRelease(
         nature=nature,
-        branch=get_branch(version),
         version=version,
         date=date,
         link=link,
         eol=eol,
     )
-
-
-def get_branch(version: str):
-    # This is a testing kernel.
-    if "rc" in version:
-        return "testing"
-    else:
-        major, minor, *_ = version.split(".")
-        return f"{major}.{minor}"
 
 
 def get_hash(kernel: KernelRelease):
@@ -80,6 +87,17 @@ def get_hash(kernel: KernelRelease):
     return f"sha256:{hash}"
 
 
+def get_oldest_branch(kernels) -> Version:
+    return min(parse_version(v) for v in kernels.keys() if v != "testing")
+
+
+def predates_oldest_branch(oldest: Version, to_compare: str) -> bool:
+    if to_compare == "testing":
+        return False
+
+    return parse_version(to_compare) < oldest
+
+
 def commit(message):
     return subprocess.check_call(["git", "commit", "-m", message, VERSIONS_FILE])
 
@@ -94,11 +112,15 @@ def main():
         sys.exit(1)
 
     releases = release_table.find_all("tr")
-    parsed_releases = filter(None, [parse_release(release) for release in releases])
+    parsed_releases = [
+        parsed for release in releases
+        if (parsed := parse_release(release)) is not None
+    ]
     all_kernels = json.load(VERSIONS_FILE.open())
+    oldest_branch = get_oldest_branch(all_kernels)
 
-    for kernel in parsed_releases:
-        branch = get_branch(kernel.version)
+    for (branch, kernels) in groupby(parsed_releases, lambda kernel: kernel.branch):
+        kernel = max(kernels, key=lambda kernel: kernel.parsed_version)
         nixpkgs_branch = branch.replace(".", "_")
 
         old_version = all_kernels.get(branch, {}).get("version")
@@ -106,7 +128,21 @@ def main():
             print(f"linux_{nixpkgs_branch}: {kernel.version} is latest, skipping...")
             continue
 
+        if predates_oldest_branch(oldest_branch, kernel.branch):
+            print(
+                f"{kernel.branch} is too old and not supported anymore, skipping...",
+                file=sys.stderr
+            )
+            continue
+
         if old_version is None:
+            if kernel.eol:
+                print(
+                    f"{kernel.branch} is EOL, not adding...",
+                    file=sys.stderr
+                )
+                continue
+
             message = f"linux_{nixpkgs_branch}: init at {kernel.version}"
         else:
             message = f"linux_{nixpkgs_branch}: {old_version} -> {kernel.version}"
